requirements.txt
Flask==2.3.3
mysql-connector-python==8.3.0
google-generativeai
Flask-Mail==0.9.1
python-dotenv
